scripts/docker-wlist/requirements.txt
boto3==1.34.149
kafka_python==2.0.2
python_logging_loki==0.3.1
Requests==2.32.3
